Fix bug in BVH This sped up rendering by about 3x - from ~35s to 11s for a test frame.
The previous code had two bugs:

1)  mixed up min and max, in a way that made the structure
very inefficient.

The goal was to see if three intervals intersected. The bug meant this
function always returned true (it essentially asked if the union of the
intervals was non-zero, which of course it was).

2) Didn't use t_max and t_min.

I saw the compiler warnings, but I mixed them up with ones ignoring t1
and t2 in a different situation (which were legit unused). A lesson to
clean up your compiler warnings!